Add 5-second delay between auto-join channels.
The time break prevents chained joining to all channels at nearly the
same time on servers where it is not common to have the invisible mode.
There, all the joined users' idents and IPs are collected with a WHO
command, resulting in many lines added to the logging files at once.